git-svn-id: svn://svn.forge.objectweb.org/svnroot/spagobi@5920 99afaf0d-6903-0410-885a-c66a8bbb5f81

+++ V2.x/trunk/SpagoBIProject/src/it/eng/spagobi/behaviouralmodel/analyticaldriver/dao/ParameterUseDAOHibImpl.java
@@ -168,7 +168,7 @@ public class ParameterUseDAOHibImpl extends AbstractHibernateDAO implements
 		     "spd.id.sbiExtRoles.name=? ";
 			
 			Query query = aSession.createQuery(hql);
-			query.setInteger(0, parameterId);
+			query.setInteger(0, parameterId.intValue());
 			query.setString(1, roleName);
 			
 			SbiParuse hibParuse = (SbiParuse)query.uniqueResult();
